Make CcRUD of categories

+++ database/migrations/2025_04_01_055531_add_category_id_to_products_table.php
@@ -11,9 +11,12 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->foreignId('category_id')->constrained()->onDelete('cascade');
-        });
+        if (!Schema::hasColumn('products', 'category_id')) {
+            Schema::table('products', function (Blueprint $table) {
+                $table->foreignId('category_id')->constrained()->onDelete('cascade');
+            });
+        }
+
     }
 
     /**
